storage/ranking_calculator/requirements.txt
connectorx==0.3.3
greenlet==3.1.1
numpy==2.1.1
pandas==2.2.3
polars==1.7.1
pyarrow==17.0.0
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
six==1.16.0
SQLAlchemy==2.0.35
typing_extensions==4.12.2
tzdata==2024.1
